pyproject.toml
[project]
name = "guardlens"
version = "0.1.0"
description = "GuardianLens — on-device AI child safety monitor powered by Gemma 4 + Ollama."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "GuardianLens Team" }]

dependencies = [
    "ollama>=0.4.0",
    "mss>=9.0.1",
    "Pillow>=10.0.0",
    "gradio>=4.44.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.4.0",
    "PyYAML>=6.0.1",
    "rich>=13.7.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.6.0",
    "mypy>=1.11.0",
    "pytest>=8.3.0",
    "types-PyYAML",
]

[project.scripts]
guardlens = "run:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guardlens", "app"]
include = ["run.py"]

[tool.ruff]
line-length = 100
target-version = "py311"
src = ["src", "app"]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # bugbear
    "SIM", # simplify
    "RUF", # ruff-specific
    "ANN", # flake8-annotations (require type hints)
    "TCH", # type-checking imports
]
ignore = [
    "ANN101", # missing self type
    "ANN102", # missing cls type
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"notebooks/**" = ["ANN", "E", "F"]
"tests/**" = ["ANN"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["mss.*", "gradio.*", "ollama.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
